Moving some files/functions around -- spitting up util.py.
util.py split into a couple of files.
