+++ src/despot/__init__.py
@@ -5,7 +5,7 @@ Despot.
 from . import rulers, util
 from . import classes
 
-from .util import register
+from .util.reg import register
 from .classes import Despot
 
 register(rulers.nero, 'nero')
