+++ src/despot/util/__init__.py
@@ -1,0 +1,3 @@
+from . import (reg, walk)
+
+__all__ = ['reg', 'walk']
