+++ tests/util/test_walk.py
@@ -1,0 +1,10 @@
+import pytest
+from despot.util.walk import walkdir, walkmodule
+
+@pytest.mark.xfail
+def test_walkdir():
+    pass
+
+@pytest.mark.xfail
+def test_walkmodule():
+    pass
